Implement new uranium ore asteroid features
They will now generate with new WeatherDef "Toxic Outer Space", which is currently a cookie cutter copy of the "Outer Space" Def - this is so that it can extend off of these same Universum features. Additionally, I have also implemented a stub WeatherEvent_CompletelyIrradiated class for future shenanigans.

They will also utilize a new, custom GenStep that enforces a new, custom (currently) permanent GameCondition "Completely Irradiated" - this is based off of the vanilla GameCondition_ToxicFallout.

I chose to do this over just using it straight up because it was fun to implement, and also, ToxicFallout requires atmosphere to be possible.